langgraph/research_paper_chatbot/research_assistant/requirements.txt
langgraph>=0.5.0
langchain-core>=0.1.0
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
scholarly>=1.7.0
